# Block/UArray/Builder: `concat` performance update