@vincenthz 2018-01-13
seems an improvement already, but it's likely we could move to keep track of values in bytes instead of unit of `ty` for the UArray. For the `Offset` unboxing, I think looking at some other point is fine